requirements.txt
numpy>=1.13.3
alpha_vantage>=2.1.0
mpl_finance>=0.10.0
matplotlib>=2.1.1
pandas>=0.23.4
python_dateutil>=2.8.0
